GrpcFullBackEnd/gateway/requirements.txt
nameko==3.0.0rc9
grpcio==1.37.0
grpcio-tools==1.37.0
